Writers updated to re-align fields
The template is now able to reliably re-align RSZ when inserting data, removing all the old padding and adding new to make it work. Also, it is much easier to insert new RSZUserDataInfos in RE2 and DMC5 by copying and inputting to a buffer inside the RSZUserDataInfoWriter to copy it from your clipboard.