Added updateCRCs script, got updated RSZParser
The RSZ Parser can now fetch CRCs from the JSON dumps, thanks to Darkness for updating it.
A 1sc script has been added to the template to automatically update CRCs that are old/incorrect to the ones from the JSON dump, to fix old modded files for new game versions.